# core/docs/swagger.yaml
...
info:
  contact: {}
  title: AI Proxy Swagger API
  version: "1.0"
paths:
...
  /mcp:
    delete:
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Host MCP Streamable Server
    get:
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Host MCP Streamable Server
    post:
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Host MCP Streamable Server
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"